make strokeWidth double

+++ Assignment3/lt-project-master/shared/src/main/scala/DSL/CanvasElementModifier.scala
@@ -12,7 +12,7 @@ case class StrokeColor(sS: String) extends CanvasElementModifier[Shape] {
   override def change(x: Shape): Unit = x.strokeColor(sS)
 }
 
-case class StrokeWidth(sW: Int) extends CanvasElementModifier[Shape] {
+case class StrokeWidth(sW: Double) extends CanvasElementModifier[Shape] {
   override def change(x: Shape): Unit = x.strokeWidth(sW)
 }
